feat: implement temperature unit preference and refactor wellness insights section
This commit introduces a user preference for temperature units (Celsius/Fahrenheit) across the application and refactors the home screen UI components for better maintainability.

- **Temperature Unit Support**:
    - Added `temperatureUnit` to `KoColorSettings` to persist user preference in DataStore.
    - Updated `HomeViewModel` to handle temperature conversion logic when Fahrenheit is selected.
    - Modified `HomeHeader` to dynamically display the correct unit symbol (°C or °F) based on current settings.
    - Integrated temperature unit selection into the `SettingsScreen` using a new `TemperatureUnitSetting` component with `FilterChip` selectors.
    - Updated `KoColorNavEntryProvider` to pass the temperature unit preference to the `Weather` route.

- **UI Refactoring**:
    - Extracted `WellnessInsightsSection`, `BioMarkerItem`, and their associated UI states from `HomeScreen.kt` into a new dedicated file: `WellnessInsightsSection.kt`.
    - Streamlined `HomeScreen.kt` by removing redundant UI logic and utilizing the new standalone components.

- **ViewModel & State Updates**:
    - Expanded `HomeUiState` and `SettingsUiState` to include temperature unit data.
    - Added `OnTempUnitChanged` event to `SettingsViewModel` to handle user preference updates.
